+++ acme/Statistics.py
@@ -39,7 +39,7 @@ class Statistics(object):
 		# retrieve or create statitics record
 		self.stats = self.setupStats()
 
-		# Start b ackground worker to handle writing to DB
+		# Start background worker to handle writing to DB
 		Logging.log('Starting statistics DB thread')
 		self.worker = BackgroundWorker.BackgroundWorker(Configuration.get('cse.statistics.writeIntervall'), self.statisticsDBWorker)
 		self.worker.start()